refactor: TextInput 컴포넌트에 Controller 적용 및 label optional로 변경
- 다른 입력 필드 컴포넌트와 동일하게 Controller 적용
- 각 입력 필드가 label 없이 사용되는 경우를 고려해 label을 optional로 변경
