+++ src/components/text-input/types.ts
@@ -1,0 +1,5 @@
+export type TextInputProps = React.InputHTMLAttributes<HTMLInputElement> & {
+  name: string;
+  label?: string;
+  required?: boolean;
+};
